Moved the image registery to own activator class
