+++ bundles/org.pitest.pitclipse.core/src/org/pitest/pitclipse/core/PitCoreActivator.java
@@ -32,29 +32,20 @@ import static org.pitest.pitclipse.runner.util.PitFileUtils.createParentDirs;
 import java.io.File;
 import java.io.IOException;
 import java.math.BigDecimal;
-import java.net.URL;
 import java.nio.file.Files;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
 
-import org.eclipse.core.runtime.FileLocator;
 import org.eclipse.core.runtime.IPath;
 import org.eclipse.core.runtime.IStatus;
-import org.eclipse.core.runtime.Path;
 import org.eclipse.core.runtime.Platform;
 import org.eclipse.core.runtime.Plugin;
 import org.eclipse.core.runtime.Status;
 import org.eclipse.core.runtime.preferences.InstanceScope;
 import org.eclipse.jface.preference.IPreferenceStore;
-import org.eclipse.jface.resource.ImageDescriptor;
-import org.eclipse.jface.resource.ImageRegistry;
-import org.eclipse.swt.graphics.Image;
-import org.eclipse.swt.widgets.Display;
 import org.eclipse.ui.preferences.ScopedPreferenceStore;
-import org.osgi.framework.Bundle;
 import org.osgi.framework.BundleContext;
-import org.osgi.framework.FrameworkUtil;
 import org.pitest.pitclipse.runner.config.PitConfiguration;
 import org.pitest.pitclipse.runner.config.PitExecutionMode;
 
@@ -88,15 +79,6 @@ public class PitCoreActivator extends Plugin {
     // The shared instance
     private static PitCoreActivator plugin;
     
-    /**
-     * Image registry for this plug in
-     */
-    private ImageRegistry imageRegistry;
-    /**
-     * Key under which the pit icon is put in the registry
-     */
-    private static final String PIT_ICON = "org.pitest.pitclipse.pitIcon";
-
     private IPreferenceStore preferences;
 
     private List<String> pitClasspath = new ArrayList<>();
@@ -106,24 +88,6 @@ public class PitCoreActivator extends Plugin {
     private File resultDir;
 
     private File historyFile;
-
-
-    /**
-     * <b>Must</b> run in UI thread.
-     */
-    private void initIcons() {
-        imageRegistry = new ImageRegistry();
-        Bundle bundle = FrameworkUtil.getBundle(getClass());
-        URL url = FileLocator.find(bundle, new Path("icons/pit.gif"), null);
-        imageRegistry.put(PIT_ICON, ImageDescriptor.createFromURL(url).createImage());
-    }
-
-    /**
-     * @return Returns the pit icon.
-     */
-    public Image getPitIcon() {
-        return imageRegistry.get(PIT_ICON);
-    }
 
     public List<String> getPitClasspath() {
         return pitClasspath;
@@ -178,8 +142,6 @@ public class PitCoreActivator extends Plugin {
             pitestJunit5PluginClasspath = new ArrayList<>();
             addMavenJarToClasspath(pitestJunit5PluginClasspath, ORG_PITEST_JUNIT5_PLUGIN, "pitest-junit5-plugin.jar");
         }
-        // needs to run in UI thread to create icons
-        Display.getDefault().syncExec(this::initIcons);
     }
 
     private void addMavenJarToClasspath(List<String> classpath, String bundleName, String jarFile) throws IOException {
