+++ bundles/org.pitest.pitclipse.ui/META-INF/MANIFEST.MF
@@ -3,6 +3,7 @@ Bundle-ManifestVersion: 2
 Bundle-Name: Pitclipse UI
 Bundle-SymbolicName: org.pitest.pitclipse.ui;singleton:=true
 Bundle-Version: 2.1.2.qualifier
+Bundle-Activator: org.pitest.pitclipse.ui.core.PitUiActivator
 Require-Bundle: org.eclipse.ui;bundle-version="[3.107.0,4.0.0)",
  org.eclipse.core.runtime;bundle-version="[3.11.1,4.0.0)",
  org.eclipse.core.resources;bundle-version="[3.10.1,4.0.0)",
@@ -18,6 +19,7 @@ Bundle-ActivationPolicy: lazy
 Bundle-RequiredExecutionEnvironment: JavaSE-1.8
 Bundle-Vendor: Pitest.org
 Bundle-ClassPath: .
-Export-Package: org.pitest.pitclipse.ui.extension.point,
+Export-Package: org.pitest.pitclipse.ui.core,
+ org.pitest.pitclipse.ui.extension.point,
  org.pitest.pitclipse.ui.utils
 Automatic-Module-Name: org.pitest.pitclipse.ui
